+++ src/translate-extensions.ts
@@ -1,2 +1,4 @@
 export {ExternalTranslationLoaderOptions} from './common/externalTranslationLoaderOptions';
 export {ExternalTranslationLoader} from './common/externalTranslationLoader';
+export {NgxTranslateStringLocalizationService} from './services/ngxTranslateStringLocalization.service';
+export {translateString} from './misc/utils';